requirements.txt
asgiref==3.8.1
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.6
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
smmap==5.0.2
SQLAlchemy==2.0.40
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
